Moved /asset folder to /framework/asset
