+++ src/framework/asset/asset-file.js
