+++ src/framework/asset/asset-reference.js
