+++ src/framework/asset/constants.js
