+++ src/framework/components/animation/component.js
@@ -7,7 +7,7 @@ import { DefaultAnimBinder } from '../../../anim/binder/default-anim-binder.js';
 
 import { Skeleton } from '../../../animation/skeleton.js';
 
-import { Asset } from '../../../asset/asset.js';
+import { Asset } from '../../asset/asset.js';
 
 import { Component } from '../component.js';
