+++ src/framework/components/collision/component.js
@@ -1,4 +1,4 @@
-import { Asset } from '../../../asset/asset.js';
+import { Asset } from '../../asset/asset.js';
 
 import { Component } from '../component.js';
