+++ src/framework/components/light/component.js
@@ -12,7 +12,7 @@ import {
     SHADOWUPDATE_REALTIME
 } from '../../../scene/constants.js';
 
-import { Asset } from '../../../asset/asset.js';
+import { Asset } from '../../asset/asset.js';
 
 import { Component } from '../component.js';
