+++ src/framework/components/particle-system/system.js
@@ -4,7 +4,7 @@ import { Vec3 } from '../../../core/math/vec3.js';
 
 import { LIGHTTYPE_DIRECTIONAL } from '../../../scene/constants.js';
 
-import { Asset } from '../../../asset/asset.js';
+import { Asset } from '../../asset/asset.js';
 
 import { Component } from '../component.js';
 import { ComponentSystem } from '../system.js';
