+++ src/framework/i18n/i18n.js
@@ -1,6 +1,6 @@
 import { EventHandler } from '../../core/event-handler.js';
 
-import { Asset } from '../../asset/asset.js';
+import { Asset } from '../asset/asset.js';
 
 import { I18nParser } from './i18n-parser.js';
